requirements.txt
bleach==2.1.4
boto==2.49.0
boto3==1.9.99
botocore==1.12.99
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.38.0
django-authtools==1.6.0
django-classy-tags==0.8.0
django-js-asset==1.2.1
django-mptt==0.9.1
django-nyt==1.1.2
django-rest-auth==0.9.3
django-sekizai==0.10.0
django-storages==1.7.1
djangorestframework==3.9.1
docutils==0.14
html5lib==1.0.1
idna==2.8
jmespath==0.9.3
Markdown==3.0.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==5.4.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sorl-thumbnail==12.5.0
urllib3==1.24.1
webencodings==0.5.1
